vttablet: move functions to state_manager.go
