+++ go/vt/vttablet/tabletserver/tabletserver.go
@@ -40,7 +40,6 @@ import (
 	"vitess.io/vitess/go/trace"
 	"vitess.io/vitess/go/vt/callerid"
 	"vitess.io/vitess/go/vt/dbconfigs"
-	"vitess.io/vitess/go/vt/dbconnpool"
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/logutil"
 	binlogdatapb "vitess.io/vitess/go/vt/proto/binlogdata"
@@ -65,59 +64,10 @@ import (
 	"vitess.io/vitess/go/vt/vttablet/tabletserver/vstreamer"
 )
 
-const (
-	// StateNotConnected is the state where tabletserver is not
-	// connected to an underlying mysql instance.
-	StateNotConnected = iota
-	// StateNotServing is the state where tabletserver is connected
-	// to an underlying mysql instance, but is not serving queries.
-	StateNotServing
-	// StateServing is where queries are allowed.
-	StateServing
-	// StateTransitioning is a transient state indicating that
-	// the tabletserver is tranisitioning to a new state.
-	// In order to achieve clean transitions, no requests are
-	// allowed during this state.
-	StateTransitioning
-	// StateShuttingDown indicates that the tabletserver
-	// is shutting down. In this state, we wait for outstanding
-	// requests and transactions to conclude.
-	StateShuttingDown
-)
-
 // logPoolFull is for throttling transaction / query pool full messages in the log.
 var logPoolFull = logutil.NewThrottledLogger("PoolFull", 1*time.Minute)
 
 var logComputeRowSerializerKey = logutil.NewThrottledLogger("ComputeRowSerializerKey", 1*time.Minute)
-
-// stateName names every state. The number of elements must
-// match the number of states. Names can overlap.
-var stateName = []string{
-	"NOT_SERVING",
-	"NOT_SERVING",
-	"SERVING",
-	"NOT_SERVING",
-	"SHUTTING_DOWN",
-}
-
-// stateDetail matches every state and optionally more information about the reason
-// why the state is serving / not serving.
-var stateDetail = []string{
-	"Not Connected",
-	"Not Serving",
-	"",
-	"Transitioning",
-	"Shutting Down",
-}
-
-// stateInfo returns a string representation of the state and optional detail
-// about the reason for the state transition
-func stateInfo(state int64) string {
-	if state == StateServing {
-		return "SERVING"
-	}
-	return fmt.Sprintf("%s (%s)", stateName[state], stateDetail[state])
-}
 
 // TabletServer implements the RPC interface for the query service.
 // TabletServer is initialized in the following sequence:
@@ -259,16 +209,23 @@ func NewTabletServer(name string, config *tabletenv.TabletConfig, topoServer *to
 	return tsv
 }
 
-// SetTracking forces tracking to be on or off.
-// Only to be used for testing.
-func (tsv *TabletServer) SetTracking(enabled bool) {
-	tsv.tracker.Enable(enabled)
-}
+// InitDBConfig initializes the db config variables for TabletServer. You must call this function before
+// calling SetServingType.
+func (tsv *TabletServer) InitDBConfig(target querypb.Target, dbcfgs *dbconfigs.DBConfigs) error {
+	tsv.mu.Lock()
+	defer tsv.mu.Unlock()
+	if tsv.state != StateNotConnected {
+		return vterrors.Errorf(vtrpcpb.Code_UNKNOWN, "InitDBConfig failed, current state: %s", stateName[tsv.state])
+	}
+	tsv.target = target
+	tsv.config.DB = dbcfgs
 
-// EnableHistorian forces historian to be on or off.
-// Only to be used for testing.
-func (tsv *TabletServer) EnableHistorian(enabled bool) {
-	_ = tsv.se.EnableHistorian(enabled)
+	tsv.se.InitDBConfig(tsv.config.DB.DbaWithDB())
+	tsv.hw.InitDBConfig(target)
+	tsv.hr.InitDBConfig(target)
+	tsv.txThrottler.InitDBConfig(target)
+	tsv.vstreamer.InitDBConfig(target.Keyspace)
+	return nil
 }
 
 // Register prepares TabletServer for serving by calling
@@ -322,60 +279,6 @@ func (tsv *TabletServer) SetQueryRules(ruleSource string, qrs *rules.Rules) erro
 	return nil
 }
 
-// GetState returns the name of the current TabletServer state.
-func (tsv *TabletServer) GetState() string {
-	if tsv.lameduck.Get() != 0 {
-		return "NOT_SERVING"
-	}
-	tsv.mu.Lock()
-	name := stateName[tsv.state]
-	tsv.mu.Unlock()
-	return name
-}
-
-// setState changes the state and logs the event.
-// It requires the caller to hold a lock on mu.
-func (tsv *TabletServer) setState(state int64) {
-	log.Infof("TabletServer state: %s -> %s", stateInfo(tsv.state), stateInfo(state))
-	tsv.state = state
-	tsv.history.Add(&historyRecord{
-		Time:         time.Now(),
-		ServingState: stateInfo(state),
-		TabletType:   tsv.target.TabletType.String(),
-	})
-}
-
-// transition obtains a lock and changes the state.
-func (tsv *TabletServer) transition(newState int64) {
-	tsv.mu.Lock()
-	tsv.setState(newState)
-	tsv.mu.Unlock()
-}
-
-// IsServing returns true if TabletServer is in SERVING state.
-func (tsv *TabletServer) IsServing() bool {
-	return tsv.GetState() == "SERVING"
-}
-
-// InitDBConfig initializes the db config variables for TabletServer. You must call this function before
-// calling SetServingType.
-func (tsv *TabletServer) InitDBConfig(target querypb.Target, dbcfgs *dbconfigs.DBConfigs) error {
-	tsv.mu.Lock()
-	defer tsv.mu.Unlock()
-	if tsv.state != StateNotConnected {
-		return vterrors.Errorf(vtrpcpb.Code_UNKNOWN, "InitDBConfig failed, current state: %s", stateName[tsv.state])
-	}
-	tsv.target = target
-	tsv.config.DB = dbcfgs
-
-	tsv.se.InitDBConfig(tsv.config.DB.DbaWithDB())
-	tsv.hw.InitDBConfig(target)
-	tsv.hr.InitDBConfig(target)
-	tsv.txThrottler.InitDBConfig(target)
-	tsv.vstreamer.InitDBConfig(target.Keyspace)
-	return nil
-}
-
 func (tsv *TabletServer) initACL(tableACLConfigFile string, enforceTableACLConfig bool) {
 	// tabletacl.Init loads ACL from file if *tableACLConfig is not empty
 	err := tableacl.Init(
@@ -414,237 +317,6 @@ func (tsv *TabletServer) InitACL(tableACLConfigFile string, enforceTableACLConfi
 	}
 }
 
-// StartService is a convenience function for InitDBConfig->SetServingType
-// with serving=true.
-func (tsv *TabletServer) StartService(target querypb.Target, dbcfgs *dbconfigs.DBConfigs) (err error) {
-	err = tsv.InitDBConfig(target, dbcfgs)
-	if err != nil {
-		return err
-	}
-	_ /* state changed */, err = tsv.SetServingType(target.TabletType, true, nil)
-	return err
-}
-
-// EnterLameduck causes tabletserver to enter the lameduck state. This
-// state causes health checks to fail, but the behavior of tabletserver
-// otherwise remains the same. Any subsequent calls to SetServingType will
-// cause the tabletserver to exit this mode.
-func (tsv *TabletServer) EnterLameduck() {
-	tsv.lameduck.Set(1)
-}
-
-// ExitLameduck causes the tabletserver to exit the lameduck mode.
-func (tsv *TabletServer) ExitLameduck() {
-	tsv.lameduck.Set(0)
-}
-
-const (
-	actionNone = iota
-	actionFullStart
-	actionServeNewType
-	actionGracefulStop
-)
-
-// SetServingType changes the serving type of the tabletserver. It starts or
-// stops internal services as deemed necessary. The tabletType determines the
-// primary serving type, while alsoAllow specifies other tablet types that
-// should also be honored for serving.
-// Returns true if the state of QueryService or the tablet type changed.
-func (tsv *TabletServer) SetServingType(tabletType topodatapb.TabletType, serving bool, alsoAllow []topodatapb.TabletType) (stateChanged bool, err error) {
-	defer tsv.ExitLameduck()
-
-	action, err := tsv.decideAction(tabletType, serving, alsoAllow)
-	if err != nil {
-		return false, err
-	}
-	switch action {
-	case actionNone:
-		return false, nil
-	case actionFullStart:
-		if err := tsv.fullStart(); err != nil {
-			tsv.closeAll()
-			return true, err
-		}
-		return true, nil
-	case actionServeNewType:
-		if err := tsv.serveNewType(); err != nil {
-			tsv.closeAll()
-			return true, err
-		}
-		return true, nil
-	case actionGracefulStop:
-		tsv.gracefulStop()
-		return true, nil
-	}
-	panic("unreachable")
-}
-
-func (tsv *TabletServer) decideAction(tabletType topodatapb.TabletType, serving bool, alsoAllow []topodatapb.TabletType) (action int, err error) {
-	tsv.mu.Lock()
-	defer tsv.mu.Unlock()
-
-	tsv.alsoAllow = alsoAllow
-
-	// Handle the case where the requested TabletType and serving state
-	// match our current state. This avoids an unnecessary transition.
-	// There's no similar shortcut if serving is false, because there
-	// are different 'not serving' states that require different actions.
-	if tsv.target.TabletType == tabletType {
-		if serving && tsv.state == StateServing {
-			// We're already in the desired state.
-			return actionNone, nil
-		}
-	}
-	tsv.target.TabletType = tabletType
-	switch tsv.state {
-	case StateNotConnected:
-		if serving {
-			tsv.setState(StateTransitioning)
-			return actionFullStart, nil
-		}
-	case StateNotServing:
-		if serving {
-			tsv.setState(StateTransitioning)
-			return actionServeNewType, nil
-		}
-	case StateServing:
-		if !serving {
-			tsv.setState(StateShuttingDown)
-			return actionGracefulStop, nil
-		}
-		tsv.setState(StateTransitioning)
-		return actionServeNewType, nil
-	case StateTransitioning, StateShuttingDown:
-		return actionNone, vterrors.Errorf(vtrpcpb.Code_FAILED_PRECONDITION, "cannot SetServingType, current state: %s", stateName[tsv.state])
-	default:
-		panic("unreachable")
-	}
-	return actionNone, nil
-}
-
-func (tsv *TabletServer) fullStart() (err error) {
-	c, err := dbconnpool.NewDBConnection(context.TODO(), tsv.config.DB.AppWithDB())
-	if err != nil {
-		log.Errorf("error creating db app connection: %v", err)
-		return err
-	}
-	c.Close()
-
-	if err := tsv.se.Open(); err != nil {
-		return err
-	}
-	tsv.vstreamer.Open()
-	if err := tsv.qe.Open(); err != nil {
-		return err
-	}
-	if err := tsv.txThrottler.Open(); err != nil {
-		return err
-	}
-	return tsv.serveNewType()
-}
-
-func (tsv *TabletServer) serveNewType() (err error) {
-	if tsv.target.TabletType == topodatapb.TabletType_MASTER {
-		tsv.watcher.Close()
-		tsv.hr.Close()
-
-		tsv.hw.Open()
-		tsv.tracker.Open()
-		tsv.te.AcceptReadWrite()
-		tsv.messager.Open()
-	} else {
-		tsv.messager.Close()
-		tsv.te.AcceptReadOnly()
-		tsv.tracker.Close()
-		tsv.hw.Close()
-		tsv.se.MakeNonMaster()
-
-		tsv.hr.Open()
-		tsv.watcher.Open()
-	}
-	tsv.transition(StateServing)
-	return nil
-}
-
-func (tsv *TabletServer) gracefulStop() {
-	defer close(tsv.setTimeBomb())
-	tsv.waitForShutdown()
-	tsv.transition(StateNotServing)
-}
-
-// StopService shuts down the tabletserver to the uninitialized state.
-// It first transitions to StateShuttingDown, then waits for active
-// services to shut down. Then it shuts down the rest. This function
-// should be called before process termination, or if MySQL is unreachable.
-// Under normal circumstances, SetServingType should be called.
-func (tsv *TabletServer) StopService() {
-	defer close(tsv.setTimeBomb())
-	defer tsv.LogError()
-
-	tsv.mu.Lock()
-	if tsv.state != StateServing && tsv.state != StateNotServing {
-		tsv.mu.Unlock()
-		return
-	}
-	tsv.setState(StateShuttingDown)
-	tsv.mu.Unlock()
-
-	log.Info("Executing complete shutdown.")
-	tsv.waitForShutdown()
-	tsv.qe.Close()
-	tsv.watcher.Close()
-	tsv.vstreamer.Close()
-	tsv.hr.Close()
-	tsv.hw.Close()
-	tsv.se.Close()
-	log.Info("Shutdown complete.")
-	tsv.transition(StateNotConnected)
-}
-
-func (tsv *TabletServer) waitForShutdown() {
-	tsv.messager.Close()
-	tsv.te.Close()
-	tsv.txThrottler.Close()
-	tsv.tracker.Close()
-	tsv.qe.StopServing()
-	tsv.requests.Wait()
-}
-
-// closeAll is called if TabletServer fails to start.
-// It forcibly shuts down everything.
-func (tsv *TabletServer) closeAll() {
-	tsv.messager.Close()
-	tsv.te.Close()
-	tsv.txThrottler.Close()
-	tsv.qe.StopServing()
-	tsv.qe.Close()
-	tsv.watcher.Close()
-	tsv.tracker.Close()
-	tsv.vstreamer.Close()
-	tsv.hr.Close()
-	tsv.hw.Close()
-	tsv.se.Close()
-	tsv.transition(StateNotConnected)
-}
-
-func (tsv *TabletServer) setTimeBomb() chan struct{} {
-	done := make(chan struct{})
-	go func() {
-		qt := tsv.QueryTimeout.Get()
-		if qt == 0 {
-			return
-		}
-		tmr := time.NewTimer(10 * qt)
-		defer tmr.Stop()
-		select {
-		case <-tmr.C:
-			log.Fatal("Shutdown took too long. Crashing")
-		case <-done:
-		}
-	}()
-	return done
-}
-
 // IsHealthy returns nil for non-serving types or if the query service is healthy (able to
 // connect to the database and serving traffic), or an error explaining
 // the unhealthiness otherwise.
@@ -667,54 +339,6 @@ func (tsv *TabletServer) IsHealthy() error {
 	default:
 		return nil
 	}
-}
-
-// CheckMySQL initiates a check to see if MySQL is reachable.
-// If not, it shuts down the query service. The check is rate-limited
-// to no more than once per second.
-// The function satisfies tabletenv.Env.
-func (tsv *TabletServer) CheckMySQL() {
-	if !tsv.checkMySQLThrottler.TryAcquire() {
-		return
-	}
-	go func() {
-		defer func() {
-			tsv.LogError()
-			time.Sleep(1 * time.Second)
-			tsv.checkMySQLThrottler.Release()
-		}()
-		if tsv.isMySQLReachable() {
-			return
-		}
-		log.Info("Check MySQL failed. Shutting down query service")
-		tsv.StopService()
-	}()
-}
-
-// isMySQLReachable returns true if we can connect to MySQL.
-// The function returns false only if the query service is
-// in StateServing or StateNotServing.
-func (tsv *TabletServer) isMySQLReachable() bool {
-	tsv.mu.Lock()
-	switch tsv.state {
-	case StateServing:
-		// Prevent transition out of this state by
-		// reserving a request.
-		tsv.requests.Add(1)
-		defer tsv.requests.Done()
-	case StateNotServing:
-		// Prevent transition out of this state by
-		// temporarily switching to StateTransitioning.
-		tsv.setState(StateTransitioning)
-		defer func() {
-			tsv.transition(StateNotServing)
-		}()
-	default:
-		tsv.mu.Unlock()
-		return true
-	}
-	tsv.mu.Unlock()
-	return tsv.qe.IsMySQLReachable()
 }
 
 // ReloadSchema reloads the schema.
@@ -1508,32 +1132,6 @@ func (tsv *TabletServer) execRequest(
 	return nil
 }
 
-// verifyTarget allows requests to be executed even in non-serving state.
-func (tsv *TabletServer) verifyTarget(ctx context.Context, target *querypb.Target) error {
-	tsv.mu.Lock()
-	defer tsv.mu.Unlock()
-
-	if target != nil {
-		// a valid target needs to be used
-		switch {
-		case target.Keyspace != tsv.target.Keyspace:
-			return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "invalid keyspace %v", target.Keyspace)
-		case target.Shard != tsv.target.Shard:
-			return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "invalid shard %v", target.Shard)
-		case target.TabletType != tsv.target.TabletType:
-			for _, otherType := range tsv.alsoAllow {
-				if target.TabletType == otherType {
-					return nil
-				}
-			}
-			return vterrors.Errorf(vtrpcpb.Code_FAILED_PRECONDITION, "invalid tablet type: %v, want: %v or %v", target.TabletType, tsv.target.TabletType, tsv.alsoAllow)
-		}
-	} else if !tabletenv.IsLocalContext(ctx) {
-		return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "No target")
-	}
-	return nil
-}
-
 func (tsv *TabletServer) handlePanicAndSendLogStats(
 	sql string,
 	bindVariables map[string]*querypb.BindVariable,
@@ -1852,52 +1450,6 @@ func (tsv *TabletServer) Close(ctx context.Context) error {
 	return nil
 }
 
-// startRequest validates the current state and target and registers
-// the request (a waitgroup) as started. Every startRequest requires
-// one and only one corresponding endRequest. When the service shuts
-// down, StopService will wait on this waitgroup to ensure that there
-// are no requests in flight.
-func (tsv *TabletServer) startRequest(ctx context.Context, target *querypb.Target, allowOnShutdown bool) (err error) {
-	tsv.mu.Lock()
-	defer tsv.mu.Unlock()
-	if tsv.state == StateServing {
-		goto verifyTarget
-	}
-	if allowOnShutdown && tsv.state == StateShuttingDown {
-		goto verifyTarget
-	}
-	return vterrors.Errorf(vtrpcpb.Code_FAILED_PRECONDITION, "operation not allowed in state %s", stateName[tsv.state])
-
-verifyTarget:
-	if target != nil {
-		// a valid target needs to be used
-		switch {
-		case target.Keyspace != tsv.target.Keyspace:
-			return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "invalid keyspace %v", target.Keyspace)
-		case target.Shard != tsv.target.Shard:
-			return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "invalid shard %v", target.Shard)
-		case target.TabletType != tsv.target.TabletType:
-			for _, otherType := range tsv.alsoAllow {
-				if target.TabletType == otherType {
-					goto ok
-				}
-			}
-			return vterrors.Errorf(vtrpcpb.Code_FAILED_PRECONDITION, "invalid tablet type: %v, want: %v or %v", target.TabletType, tsv.target.TabletType, tsv.alsoAllow)
-		}
-	} else if !tabletenv.IsLocalContext(ctx) {
-		return vterrors.Errorf(vtrpcpb.Code_INVALID_ARGUMENT, "No target")
-	}
-
-ok:
-	tsv.requests.Add(1)
-	return nil
-}
-
-// endRequest unregisters the current request (a waitgroup) as done.
-func (tsv *TabletServer) endRequest() {
-	tsv.requests.Done()
-}
-
 func (tsv *TabletServer) registerDebugHealthHandler() {
 	tsv.exporter.HandleFunc("/debug/health", func(w http.ResponseWriter, r *http.Request) {
 		if err := acl.CheckAccessHTTP(r, acl.MONITORING); err != nil {
@@ -1938,6 +1490,18 @@ func (tsv *TabletServer) registerTwopczHandler() {
 		}
 		twopczHandler(txe, w, r)
 	})
+}
+
+// SetTracking forces tracking to be on or off.
+// Only to be used for testing.
+func (tsv *TabletServer) SetTracking(enabled bool) {
+	tsv.tracker.Enable(enabled)
+}
+
+// EnableHistorian forces historian to be on or off.
+// Only to be used for testing.
+func (tsv *TabletServer) EnableHistorian(enabled bool) {
+	_ = tsv.se.EnableHistorian(enabled)
 }
 
 // SetPoolSize changes the pool size to the specified value.
